ipg2lin

Takes the resulting file of an efetch run on the ipg database and

Takes the resulting file of an efetch run on the ipg database and append lineage, and taxid columns

Usage

ipg2lin(
  accessions,
  ipg_file,
  refseq_assembly_path,
  genbank_assembly_path,
  lineagelookup_path
)

ipg2lin(
  accessions,
  ipg_file,
  refseq_assembly_path,
  genbank_assembly_path,
  lineagelookup_path
)

Arguments

accessions

Character vector of protein accessions

ipg_file

Path to the file containing results of an efetch run on the ipg database. The protein accession in 'accessions' should be contained in this file

lineagelookup_path

String of the path to the lineage lookup file (taxid to lineage mapping). This file can be generated using the "create_lineage_lookup()" function

assembly_path

String of the path to the assembly_summary path This file can be generated using the "DownloadAssemblySummary()" function

Value

Describe return, in detail

Author

Samuel Chen, Janani Ravi

Examples

if (FALSE) { # \dontrun{
ipg2lin()
} # }