Given a pin_id, returns the URL where the user can check the status of their job

The 'base_url' parameter is set from the environment variable 'BASE_URL' if it's available, and defaults to "http://jravilab.org/molevolvr/" if it's not.

Usage

make_job_results_url(
  pin_id,
  base_url = Sys.getenv("BASE_URL", unset = "http://jravilab.org/molevolvr/")
)

Arguments

pin_id

the unique identifier for the job

base_url

the base URL for the MolEvolvR web app

Value

the URL where the user can check the status of their job